+++ angular/src/shared/shared.module.ts
@@ -8,13 +8,13 @@ import { AppSessionService } from './session/app-session.service';
 import { AppUrlService } from './nav/app-url.service';
 import { AppAuthService } from './auth/app-auth.service';
 import { AppRouteGuard } from './auth/auth-route-guard';
-import { AbpPaginationControlsComponent } from './pagination/abp-pagination-controls.component';
 import { LocalizePipe } from '@shared/pipes/localize.pipe';
 import { DragDropModule } from '@angular/cdk/drag-drop';
 import { ScrollingModule } from '@angular/cdk/scrolling';
 import { CdkTableModule } from '@angular/cdk/table';
 import { CdkTreeModule } from '@angular/cdk/tree';
 
+import { AbpPaginationControlsComponent } from './components/pagination/abp-pagination-controls.component';
 import { ApbValidationSummaryComponent } from './components/validation/abp-validation.summary.component';
 import { AbpModalHeaderComponent } from './components/modal/abp-modal-header.component';
 import { AbpModalFooterComponent } from './components/modal/abp-modal-footer.component';
